ItemSelectionPanel in Users and Applications is missing toolbar with Clear/Show All buttons (#4449)
Moved toolbar logic and styles to common `BrowseItemsSelectionPanel` class.
