+++ modules/admin/admin-ui/src/main/resources/web/admin/apps/applications/js/app/browse/ApplicationBrowseItemPanel.ts
@@ -1,11 +1,12 @@
 import '../../api.ts';
 import {ApplicationBrowseItemsSelectionPanel} from './ApplicationBrowseItemsSelectionPanel';
 import {ApplicationItemStatisticsPanel} from '../view/ApplicationItemStatisticsPanel';
+import {ApplicationTreeGrid} from './ApplicationTreeGrid';
 
 export class ApplicationBrowseItemPanel extends api.app.browse.BrowseItemPanel<api.application.Application> {
 
-    createItemSelectionPanel(): ApplicationBrowseItemsSelectionPanel {
-        return new ApplicationBrowseItemsSelectionPanel();
+    createItemSelectionPanel(grid: ApplicationTreeGrid): ApplicationBrowseItemsSelectionPanel {
+        return new ApplicationBrowseItemsSelectionPanel(grid);
     }
 
     createItemStatisticsPanel(): api.app.view.ItemStatisticsPanel<api.application.Application> {
