+++ modules/admin/admin-ui/src/main/resources/web/admin/apps/applications/js/app/browse/ApplicationBrowseItemsSelectionPanel.ts
@@ -3,8 +3,9 @@ import '../../api.ts';
 import BrowseItem = api.app.browse.BrowseItem;
 import Application = api.application.Application;
 import ApplicationViewer = api.application.ApplicationViewer;
+import BrowseItemsSelectionPanel = api.app.browse.BrowseItemsSelectionPanel;
 
-export class ApplicationBrowseItemsSelectionPanel extends api.app.browse.BrowseItemsSelectionPanel<Application> {
+export class ApplicationBrowseItemsSelectionPanel extends BrowseItemsSelectionPanel<Application> {
 
     createItemViewer(item: BrowseItem<Application>): ApplicationViewer {
         let viewer = new ApplicationViewer();
