+++ modules/admin/admin-ui/src/main/resources/web/admin/apps/user-manager/js/app/browse/UserBrowseItemPanel.ts
@@ -2,11 +2,12 @@ import '../../api.ts';
 import {UserBrowseItemsSelectionPanel} from './UserBrowseItemsSelectionPanel';
 import {UserTreeGridItem} from './UserTreeGridItem';
 import {UserItemStatisticsPanel} from '../view/UserItemStatisticsPanel';
+import {UserItemsTreeGrid} from './UserItemsTreeGrid';
 
 export class UserBrowseItemPanel extends api.app.browse.BrowseItemPanel<UserTreeGridItem> {
 
-    createItemSelectionPanel(): UserBrowseItemsSelectionPanel {
-        return new UserBrowseItemsSelectionPanel();
+    createItemSelectionPanel(grid: UserItemsTreeGrid): UserBrowseItemsSelectionPanel {
+        return new UserBrowseItemsSelectionPanel(grid);
     }
 
     createItemStatisticsPanel(): api.app.view.ItemStatisticsPanel<UserTreeGridItem> {
