+++ modules/admin/admin-ui/src/main/resources/web/admin/apps/user-manager/js/app/browse/UserBrowseItemsSelectionPanel.ts
@@ -3,8 +3,9 @@ import {UserTreeGridItemViewer} from './UserTreeGridItemViewer';
 import {UserTreeGridItem} from './UserTreeGridItem';
 
 import BrowseItem = api.app.browse.BrowseItem;
+import BrowseItemsSelectionPanel = api.app.browse.BrowseItemsSelectionPanel;
 
-export class UserBrowseItemsSelectionPanel extends api.app.browse.BrowseItemsSelectionPanel<UserTreeGridItem> {
+export class UserBrowseItemsSelectionPanel extends BrowseItemsSelectionPanel<UserTreeGridItem> {
 
     createItemViewer(item: BrowseItem<UserTreeGridItem>): UserTreeGridItemViewer {
         let viewer = new UserTreeGridItemViewer();
