+++ modules/admin/admin-ui/src/main/resources/web/admin/apps/user-manager/js/app/browse/UserBrowsePanel.ts
@@ -53,7 +53,7 @@ export class UserBrowsePanel extends api.app.browse.BrowsePanel<UserTreeGridItem
     }
 
     protected createBrowseItemPanel(): UserBrowseItemPanel {
-        return new UserBrowseItemPanel();
+        return new UserBrowseItemPanel(this.treeGrid);
     }
 
     protected createFilterPanel(): PrincipalBrowseFilterPanel {
