+++ modules/admin/admin-ui/src/main/resources/web/admin/common/js/app/browse/action/_module.ts
@@ -1,1 +1,2 @@
 ///<reference path='ToggleFilterPanelAction.ts' />
+///<reference path='ShowAllAction.ts' />
