+++ modules/admin/admin-ui/src/main/resources/web/admin/common/styles/apps/content/browse/content-browse-items-selection-panel.less
@@ -5,42 +5,6 @@
 
   &.no-selection {
     margin-bottom: 0px;
-
-    .toolbar {
-      display: none;
-    }
   }
 
-  &.no-toolbar {
-    .show-all, .fold-button {
-      display: none !important;
-    }
-  }
-
-  .toolbar {
-    background-color: @admin-bg-almost-white;
-    border-top: 1px solid @admin-bg-light-gray;
-    text-align: center;
-
-    .@{_COMMON_PREFIX}button {
-      &:active, span {
-        color: @admin-button-blue2;
-      }
-
-      &:hover:not(:disabled) {
-        span {
-          color: @admin-button-blue3;
-        }
-      }
-
-      &:focus {
-        box-shadow: none;
-        border: none;
-      }
-    }
-
-    .@{_COMMON_PREFIX}dropdown {
-      bottom: 32px;
-    }
-  }
 }
